--- a/gemm-common/src/gemm.rs
+++ b/gemm-common/src/gemm.rs
@@ -97,7 +97,7 @@ impl Conj for c64 {
     }
 }
 
-pub const DEFAULT_THREADING_THRESHOLD: usize = 48 * 48 * 256;
+pub const DEFAULT_THREADING_THRESHOLD: usize = 64 * 64 * 256;
 pub const DEFAULT_RHS_PACKING_THRESHOLD: usize = 128;
 pub const DEFAULT_LHS_PACKING_THRESHOLD_SINGLE_THREAD: usize = 8;
 pub const DEFAULT_LHS_PACKING_THRESHOLD_MULTI_THREAD: usize = 16;
@@ -345,17 +345,17 @@ pub unsafe fn gemm_basic_generic<
 
             let n_threads = match parallelism {
                 Parallelism::None => 1,
-                Parallelism::Rayon(n_threads) => {
+                Parallelism::Rayon(max_threads) => {
                     let threading_threshold = get_threading_threshold();
-                    if m * n_chunk * k_chunk <= threading_threshold {
-                        1
+
+                    let max_threads = if n_threads == 0 {
+                        rayon::current_num_threads()
                     } else {
-                        if n_threads == 0 {
-                            rayon::current_num_threads()
-                        } else {
-                            n_threads
-                        }
-                    }
+                        max_threads
+                    };
+                    let total_work = m * n_chunk * k_chunk;
+                    let n_threads = std::cmp::max(1, std::cmp::min(max_threads, (total_work - threading_threshold + 1) / threading_threshold));
+                    n_threads
                 }
             };